requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==16.0
python-dotenv==1.0.1
google-genai==1.65.0
langchain==1.2.10
langchain-google-genai==4.2.1
google-cloud-storage==2.19.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.18
PyPDF2==3.0.1
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
slowapi==0.1.9
